requirements.txt
flask
pillow
opencv-python
numpy
scikit-learn
joblib
